+++ src/family_negative_binomial.cpp
@@ -3,7 +3,7 @@
     File: family_negative_binomial.cpp
     Purpose: Implementation of Negative Binomial family for different link functions
     Author: Steffen Maletz
-    Last modified: 2025-12-06
+    Last modified: 2026-07-14
 -----------------------------------------------------------------------------
 */
 
@@ -307,7 +307,8 @@ const bool LogNegativeBinomial::valid_link(const arma::mat &x) const
 
 /*
   Define square root model, i.e. the 'sqrt' link of the Negative Binomial distribution, i.e., link function: g(mu) = sqrt(mu)
-  Observation transformation: h(y) = 2.0 * sqrt(y + 3/8) Anscombe transformation (of Poisson case, as approximation)
+  Observation transformation: h(y) = 2.0 * sqrt(y + 3/8) Anscombe transformation (of Poisson case, as approximation) (Outdated, because caused problems)
+
   Link transformation: h(psi) = psi
 */
 
@@ -366,7 +367,8 @@ const double SqrtNegativeBinomial::derivative_inverse_link(const double x) const
 
 const double SqrtNegativeBinomial::observation_trafo(const double x) const
 {
-  return std::sqrt(x + 3.0 / 8.0) * 2.0;
+  // return std::sqrt(x + 3.0 / 8.0) * 2.0;
+  return std::sqrt(x);
 }
 
 const double SqrtNegativeBinomial::link_trafo(const double x) const